Refactor : 공통 파일들 components의 common으로 이동
