+++ histudy-front/src/components/common/ARouter.js
@@ -1,28 +1,28 @@
 import { BrowserRouter, Route, Router, Routes } from "react-router-dom";
-import Enroll from "../pages/Enroll/Enroll";
-import Group from "../pages/Group/Group";
-import Main from "../pages/Main/Main";
-import Post from "../pages/Post/Post";
-import Rank from "../pages/Rank/Rank";
-import Report from "../pages/Report/Report";
+import Enroll from "../../pages/Enroll/Enroll";
+import Group from "../../pages/Group/Group";
+import Main from "../../pages/Main/Main";
+import Post from "../../pages/Post/Post";
+import Rank from "../../pages/Rank/Rank";
+import Report from "../../pages/Report/Report";
 import Footer from "./Footer";
 import Header from "./Header";
-import ManageGroup from "../pages/Manager/ManageGroup";
-import ManageClass from "../pages/Manager/ManageClass";
-import StudyGroup from "../pages/Manager/StudyGroup";
-import ManageStudent from "../pages/Manager/ManageStudent";
-import ManageReport from "../pages/Manager/ManageReport";
-import ReportDetail from "../pages/Manager/ReportDetail";
-import CreateGroup from "../pages/Manager/CreateGroup";
-import MainTest from "./Main/MainTest";
-import Snackbars from "../pages/Manager/Snackbars";
+import ManageGroup from "../../pages/Manager/ManageGroup";
+import ManageClass from "../../pages/Manager/ManageClass";
+import StudyGroup from "../../pages/Manager/StudyGroup";
+import ManageStudent from "../../pages/Manager/ManageStudent";
+import ManageReport from "../../pages/Manager/ManageReport";
+import ReportDetail from "../../pages/Manager/ReportDetail";
+import CreateGroup from "../../pages/Manager/CreateGroup";
+import MainTest from "../Main/MainTest";
+import Snackbars from "../../pages/Manager/Snackbars";
 import { useRecoilState, useRecoilValue } from "recoil";
-import { isDelete, isLoadingState } from "../store/atom";
-import Profile from "../pages/Profile/Profile";
+import { isDelete, isLoadingState } from "../../store/atom";
+import Profile from "../../pages/Profile/Profile";
 import { motion } from "framer-motion";
 import { Box } from "@mui/material";
 import LoadingLottie from "./LoadingLottie";
-import PrivateRoute from "../auth/PrivateRoute";
+import PrivateRoute from "../../auth/PrivateRoute";
 
 export default function ARouter() {
   const isLoading = useRecoilValue(isLoadingState);
