+++ histudy-front/src/components/common/PostBox.js
