+++ histudy-front/src/pages/Manager/ManageGroup.js
@@ -8,10 +8,10 @@ import {
 import { border, Box } from "@mui/system";
 import { useEffect, useState } from "react";
 import SearchIcon from "@mui/icons-material/Search";
-import CustomTable from "../../components/CustomTable";
-import LongButton from "../../components/LongButton";
-import GrayBorderBox from "../../components/GrayBorderBox";
-import ProgressBar from "../../components/ProgressBar";
+import CustomTable from "../../components/common/CustomTable";
+import LongButton from "../../components/common/LongButton";
+import GrayBorderBox from "../../components/common/GrayBorderBox";
+import ProgressBar from "../../components/common/ProgressBar";
 import SideBar from "../../components/Manager/SideBar";
 import YearSelectButton from "../../components/Manager/YearSelectButton";
 import SemesterSelectButton from "../../components/Manager/SemesterSelectButton";
@@ -32,7 +32,7 @@ import Title from "../../components/Manager/Table/Title";
 import { StyledLayout } from "./style/StyledLatout";
 import { StyledTitleFlexBox } from "./style/StyledTitleFlexBox";
 import { useQueries, useQuery } from "react-query";
-import LoadingLottie from "../../components/LoadingLottie";
+import LoadingLottie from "../../components/common/LoadingLottie";
 import LoadingLayout from "../../components/Manager/Loading/LoadingLayout";
 
 const SizedBox = styled(Box)({
